Move test cases to advection
