+++ tests/geometryRTheta/advection_rtheta/advection_selected_test.cpp
@@ -8,6 +8,8 @@
 
 #include <ddc/ddc.hpp>
 
+#include "../../advection/r_theta_test_cases.hpp"
+
 #include "advection_simulation_utils.hpp"
 #include "bsl_advection_polar.hpp"
 #include "cartesian_to_circular.hpp"
@@ -33,7 +35,6 @@
 #include "rk3.hpp"
 #include "rk4.hpp"
 #include "spline_polar_foot_finder.hpp"
-#include "test_cases.hpp"
